+++ engine/flight/FlightIntegrator.h
@@ -14,7 +14,7 @@ namespace fl {
 
 // Aircraft state vector — flat/POD for network serialisation.
 struct FlightState {
-    float pos_world[3]{};          // world-frame position (m)
+    double pos_world[3]{};         // world-frame position (m) — double for planet-scale precision
     float vel_body[3]{};           // body-frame velocity (m/s)
     float euler[3]{};              // roll, pitch, yaw (rad) — derived from quaternion
     float omega[3]{};              // body-frame angular rates: p, q, r (rad/s)
